Remove tprint.c from link units; embed via ki-common.h
- cifar/Makefile: drop TPRINT_SRC from LDLIBS; keep only GLIB_* flags
- fashion/Makefile: drop TPRINT_SRC from LDLIBS; keep only GLIB_* flags
- mnist/Makefile: remove TPRINT_SRC from every rule and from OTTO/HEBB/ADAM/BITVOTE/VIS/MERGE targets
- lib/maj3.h: wrap three sign-conversion expressions with GCC diagnostic push/pop to silence -Werror while preserving original logic
- lib/tprint.c: change all public prototypes to static so each TU receives its own copy when included
- lib/tprint.h: add 2026-08-17 header comment explaining static embedding; mark every declaration static
- mnist/ki-common.h: insert guarded #include of ../lib/tprint.h + ../lib/tprint.c with diagnostic suppression block
- mnist/merge-ensemble.c, mlp-*-trn*.c, mlp-*-vis*.c, mlp-*-xform*.c: delete explicit ../lib/tprint.h includes (now provided by ki-common.h)
- mnist/member-*.out: add two new cached optimal-subset files for H196-E10 runs (OT8 and BV8)

Rationale: tprint.c now lives inside ki-common.h as static code, eliminating per-Makefile linking, duplicate symbols, and pkg-config ordering issues. Diagnostic pragmas keep -Werror builds clean.

Impact: single-source change; no behavioural difference; build recipes simplified; new member files speed up CI for ensemble experiments.

+++ mnist/mlp-bin32-hebbian-trn.c
@@ -19,7 +19,6 @@
 #define _POSIX_C_SOURCE 200809L
 #define KI_COMMON_LOAD_INPUT
 #include "ki-config.h"
-#include "../lib/tprint.h"   /* before ki-common.h: ki-common.h uses TPrint/tprint_* */
 #include "ki-common.h"
 #include "maj3.h"
 #include "w0_random.h"
@@ -156,7 +155,7 @@ static inline void member_score(const uint32_t *h0, const uint32_t *W1,
     for (int k = 0; k < N_CLASSES; k++) {
         int64_t sum = 0;
         for (int h = 0; h < H; h++)
-            sum += (int64_t)__builtin_popcount(~(W1[(size_t)k * (size_t)H + h] ^ h0[h]));
+            sum += (int64_t)__builtin_popcount(~(W1[(size_t)k * (size_t)H + (size_t) h] ^ h0[h]));
         scores[k] = sum;
     }
 }
